pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "duplicates-detector"
dynamic = ["version"]
description = "CLI tool for detecting duplicate/similar video files"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "rapidfuzz>=3.0,<4",
    "rich>=13.0,<15",
    "pdqhash>=0.2.2,<1",
    "Pillow>=10.0,<13",
    "tomli>=2.0,<3; python_version < '3.11'",
    "tomli_w>=1.0,<2",
    "shtab>=1.7,<2",
]

[project.optional-dependencies]
trash = ["send2trash>=1.8,<3"]
ssim = ["scikit-image>=0.21,<1"]
audio = ["mutagen>=1.47,<2"]
clip = ["onnxruntime>=1.16,<2"]
document = ["pypdf>=4.0,<6", "pdfminer.six>=20221105,<20270000", "python-docx>=1.0,<2", "scikit-learn>=1.3,<2"]
dev = [
    "argparse-manpage>=4,<5",
    "pytest>=8.0,<10",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<8",
    "send2trash>=1.8,<3",
    "scikit-image>=0.21,<1",
    "mutagen>=1.47,<2",
    "pypdf>=4.0,<6",
    "pdfminer.six>=20221105,<20270000",
    "python-docx>=1.0,<2",
    "scikit-learn>=1.3,<2",
    "ruff>=0.4,<1",
    "pre-commit>=3.0,<5",
    "pyright>=1.1,<2",
]

[project.scripts]
duplicates-detector = "duplicates_detector.cli:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
artifacts = ["duplicates_detector/_resources/*.js"]

[tool.hatch.build.targets.wheel.shared-data]
"man/duplicates-detector.1" = "share/man/man1/duplicates-detector.1"

[tool.hatch.build.hooks.vcs]
version-file = "duplicates_detector/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"'). Used by tests/test_integration.py.",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "I001",    # import sorting — baseline, fix in follow-up
    "SIM102",  # collapsible-if — intentional readability choice
    "SIM105",  # suppressible-exception — intentional clarity
    "SIM108",  # ternary — intentional readability choice
    "SIM115",  # open-file-with-context-handler — managed separately
    "SIM117",  # multiple-with-statements — valid Python pattern
    "B904",    # raise-without-from — baseline, fix in follow-up
    "B905",    # zip-without-strict — baseline, fix in follow-up
    "E701",    # multiple-statements-on-one-line — compact match arms
    "E741",    # ambiguous-variable-name — short lambda vars in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "E501", "F401", "F841"]
